Refactor, file relocate
