+++ vulkan/Image.h
@@ -37,7 +37,7 @@ public:
 	void UpdateByteStream(const GliImageWrapper& gliTex);
 	void UpdateByteStream(const GliImageWrapper& gliTex, uint32_t layer);
 
-	void CopyFromBuffer(const std::shared_ptr<Buffer>& pBuffer, const std::shared_ptr<CommandBuffer>& pCommandBuffer, bool waitUntilComplete = false);
+	void CopyFromBuffer(const std::shared_ptr<Buffer>& pBuffer, const std::shared_ptr<CommandBuffer>& pCommandBuffer);
 
 	virtual std::shared_ptr<ImageView> CreateDefaultImageView(bool isStorage = false) const;
 	virtual std::shared_ptr<ImageView> CreateImageView(uint32_t mipLevel, bool isStorage = false) const;
